// types/PDF.type.ts
export type PdfProp = {
  title: string;
  text: string;
  expertiseAt: string;
  assignatureLine: string;
  professional: string;
  crm: string;
  crmExpeditor: string;
  crmCurseNumber: string;
  crmCurse: string;
};

// <BasicDocument
//         title={title}
//         text={text}
//         expertiseAt={expertiseAt}
//         assignatureLine={line}
//         professional={professional}
//         crm={crm}
//         crmExpeditor={crmExpeditor}
//         crmCurseNumber={crmCurseNumber}
//         crmCurse={crmCurse}
//       />